requirements.txt
appdirs==1.4.2
asgiref==1.0.0
attrs==16.3.0
autobahn==0.17.2
Automat==0.5.0
channels==1.0.3
constantly==15.1.0
daphne==1.0.3
Django==1.10.6
djangorestframework==3.5.4
djangorestframework-jwt==1.9.0
httplib2==0.10.3
incremental==16.10.1
Jinja2==2.9.5
MarkupSafe==0.23
numpy==1.12.0
oauth2client==4.0.0
packaging==16.8
phonenumbers==8.3.1
pyasn1==0.2.3
pyasn1-modules==0.0.8
PyJWT==1.4.2
pyparsing==2.1.10
requests==2.13.0
rsa==3.4.2
six==1.10.0
Twisted==17.1.0
txaio==2.6.1
zope.interface==4.3.3
